WV-102
Reserved characters are defined in RFC 3986:
http://tools.ietf.org/html/rfc3986#section-2.2

encodeURIComponent is used and escapes all reserved characters except those
that are not used in URIs:

https://developer.mozilla.org/en-US/docs/JavaScript/Reference/Global_Objects/encodeURIComponent

For this to work properly in the current structure, each component must only
specify one value.

Also refactored the permalink code which was global into the
Worldview.Permalink namespace and split it out into several functions to make
most of it testable. Commas in the map extent were replaced with underscores,
but it can accept both when parsing. Hours/minutes/seconds were removed from
the dateSpan parameter.

Consolidated more startup code into main.js. It was confusing having some
startup code in index.html and others in the javascript itself.