Arrows replaced with colons.
Also note that the colons were smashed together without any whitespace. I added whitespace to make it look nicer, but it seems that was done to get everything to fit inside the box that contains the layer name. Since the interface is going to change anyway, keep the whitespace in and lets make it work this time around.