refactor: Remove unnecessary condition in LayerRow component
The code changes remove an unnecessary condition in the LayerRow component that checked for the 'underZoomValue' property of the 'zot' object. This condition was causing the component to be disabled even when there were granules present. Removing this condition ensures that the component is only disabled when there are no granules.