WV-3035 Global Granule Footprints (#5406)
* feat: Add CMR source and WebGL layer for displaying granules

The code changes add a new CMR source and a WebGL layer to display granules on the map. This allows for fetching and rendering granule data from the CMR API.

* refactor: Remove unnecessary condition in LayerRow component

The code changes remove an unnecessary condition in the LayerRow component that checked for the 'underZoomValue' property of the 'zot' object. This condition was causing the component to be disabled even when there were granules present. Removing this condition ensures that the component is only disabled when there are no granules.

* refactor: Use OL Vector Layer instead of WebGL Layer