chore: Update TEMPO layer configurations (#5506)
Update the TEMPO layer configurations to include the "shiftadjacentdays" property with a value of false. This property controls whether the granules should be shifted across the dateline. By setting it to false, the granules will no longer be shifted. This change ensures that the TEMPO layer configurations are aligned with the desired behavior.

Refactor granuleLayerBuilder

Refactor the granuleLayerBuilder function to include a check for the "shiftadjacentdays" property in the layer configuration definition. If the property is not present or is set to true, the granules will be shifted across the dateline as before. However, if the property is set to false, the granules will no longer be shifted. This change provides more flexibility in controlling the behavior of the granule layer builder.